Switch velocity to doubles, add per-ball friction factor
- Ball.vel is now Point2D.Double — small collision impulses no longer
  truncate to zero, so slow balls properly transfer energy
- BilliardStrategy.force changed from Point to doubles for the same reason
- frictionFactor added to Ball (default 1.0); applied each frame in
  UpdateCommand before the update strategy runs
- Straight, rotate, shrinkgrow, dejavu, straighttohell balls get 0.99
  friction; wander, mowthelawn, cornergravity, drag stay at 1.0
- MowTheLawn sweep speed randomized to 3–8 px/frame (was fixed at 4)

+++ src/main/java/edu/rice/comp504/model/strategy/updateStrategy/DejaVuStrategy.java
@@ -35,11 +35,11 @@ public class DejaVuStrategy implements IUpdateStrategy {
      * @param context The ball to update
      */
     public void updateState(Ball context) {
-        context.nextLocation((int) context.getVelocity().getX(), (int) context.getVelocity().getY());
+        context.nextLocation(context.getVelocity().getX(), context.getVelocity().getY());
         jump_clock++;
         if (jump_clock > 10) {
             Point newloc = context.getLocation();
-            newloc = new Point((int) newloc.getX() - 4 * (int) context.getVelocity().getX(), (int) newloc.getY() - 4 * (int) context.getVelocity().getY());
+            newloc = new Point((int)(newloc.getX() - 4 * context.getVelocity().getX()), (int)(newloc.getY() - 4 * context.getVelocity().getY()));
             context.setLocation(newloc);
             jump_clock = 0;
         }
